+++ src/Array/ArrayKit.php
@@ -7,8 +7,9 @@
  *
  * PHP version 8.1
  *
- * @package Inane\Stdlib
  * @author Philip Michael Raab<peep@inane.co.za>
+ * @package Inane\Stdlib
+ * @category array
  *
  * @license UNLICENSE
  * @license https://github.com/inanepain/stdlib/raw/develop/UNLICENSE UNLICENSE
@@ -21,7 +22,10 @@ declare(strict_types=1);
 
 namespace Inane\Stdlib\Array;
 
+use Inane\Stdlib\Converters\Arrayable;
+
 use function array_push;
+use function array_unshift;
 use function call_user_func;
 use function count;
 use function function_exists;
@@ -34,19 +38,19 @@ use const null;
  * Methods listed bellow have been tested, but only their simplest use case.
  *
  * @method array filter(callable $func) Filters elements this array using a callback function
- * @method array map(callable $func) Applies the callback to the elements of this array
- * @method array merge(array $array) Merges an array into this array
- * @method mixed pop() Pop the element off the end of array
- * @method mixed shift() Shift an element off the beginning of array
+ * @method array map(callable $func)    Applies the callback to the elements of this array
+ * @method array merge(array $array)    Merges an array into this array
+ * @method mixed pop()                  Pop the element off the end of array
+ * @method mixed shift()                Shift an element off the beginning of array
  * @method int   unshift(mixed $values) Prepend one or more elements to the beginning of an array
- * @method int   push(mixed $values) Push one or more elements onto the end of array
- * @method array walk(callable $func) Apply a user supplied function to every member of this array
+ * @method int   push(mixed $values)    Push one or more elements onto the end of array
+ * @method array walk(callable $func)   Apply a user supplied function to every member of this array
  *
  * @package Inane\Stdlib\Array
  *
  * @version 0.1.0
  */
-class ArrayKit {
+class ArrayKit implements Arrayable {
     /**
      * List of array functions that have the array before the callback
      *
@@ -87,6 +91,15 @@ class ArrayKit {
     }
 
     /**
+     * to array
+     *
+     * @return array
+     */
+    public function toArray(): array {
+        return $this->data;
+    }
+
+    /**
      * __invoke
      *
      * @param string $name of array method
@@ -103,7 +116,8 @@ class ArrayKit {
             else if (in_array($func, static::$before))
                 $result = $func($this->data, ...$arguments);
             else {
-                array_push($arguments, $this->data);
+                // array_push($arguments, $this->data);
+                array_unshift($arguments, $this->data);
                 $result = @call_user_func($func, ...$arguments);
             }
 
@@ -112,14 +126,5 @@ class ArrayKit {
         }
 
         return null;
-    }
-
-    /**
-     * get array
-     *
-     * @return array
-     */
-    public function getArray(): array {
-        return $this->data;
     }
 }