+++ src/ArrayObject.php
@@ -7,8 +7,9 @@
  *
  * PHP version 8.1
  *
- * @package Inane\Stdlib
  * @author  Philip Michael Raab<peep@inane.co.za>
+ * @package Inane\Stdlib
+ * @category array
  *
  * @license UNLICENSE
  * @license https://github.com/inanepain/stdlib/raw/develop/UNLICENSE UNLICENSE