+++ src/Notice/NoticeInterface.php
@@ -1,0 +1,45 @@
+<?php
+
+/**
+ * Inane: Stdlib
+ *
+ * Inane Standard Library
+ *
+ * PHP version 8.1
+ *
+ * @author Philip Michael Raab<peep@inane.co.za>
+ * @package Inane\Stdlib
+ *
+ * @license UNLICENSE
+ * @license https://github.com/inanepain/stdlib/raw/develop/UNLICENSE UNLICENSE
+ *
+ * @version $Id$
+ * $Date$
+ */
+
+/**
+ * Notice:
+ *
+ *  - E_USER_ERROR
+ *  - E_USER_WARNING
+ *  - E_USER_NOTICE
+ *  - E_USER_DEPRECATED
+ *
+ * UserNotice
+ * 	Exception thrown if a callback refers to an undefined function or if some arguments are missing.
+ */
+declare(strict_types=1);
+
+namespace Inane\Stdlib\Notice;
+
+/**
+ * NoticeInterface
+ *
+ *  - 0  : Notice
+ *
+ * @package Inane\Notice
+ *
+ * @version 0.1.0
+ */
+interface NoticeInterface {
+}