+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 0.3.0-dev @2022 Oct xx
 
 - new: `Options::empty` Test if empty
+- new: `Options::prev` Rewinds the internal pointer by 1
 - new: `Inflector::breakOnUppercase` Break on uppercase letters
 - todo: `...able` more able interfaces
 - todo: `ArrayKit` test more functions and add @method tags for them